+++ odoo/js/core/abstract/screen.js
@@ -5,6 +5,9 @@ odoo.define("terminal.core.abstract.Screen", function (require) {
     "use strict";
 
     const Class = require("web.Class");
+    const core = require("web.core");
+
+    const _t = core._t;
 
     const AbstractScreen = Class.extend({
         init: function (options) {
@@ -16,94 +19,94 @@ odoo.define("terminal.core.abstract.Screen", function (require) {
         },
 
         destroy: function () {
-            throw Error("Not Implemented!");
+            throw Error(_t("Not Implemented!"));
         },
 
         getContent: function () {
-            throw Error("Not Implemented!");
+            throw Error(_t("Not Implemented!"));
         },
 
         getBuffer: function () {
-            throw Error("Not Implemented!");
+            throw Error(_t("Not Implemented!"));
         },
 
         scrollDown: function () {
-            throw Error("Not Implemented!");
+            throw Error(_t("Not Implemented!"));
         },
 
         clean: function () {
-            throw Error("Not Implemented!");
+            throw Error(_t("Not Implemented!"));
         },
 
         cleanInput: function () {
-            throw Error("Not Implemented!");
+            throw Error(_t("Not Implemented!"));
         },
 
         cleanShadowInput: function () {
-            throw Error("Not Implemented!");
+            throw Error(_t("Not Implemented!"));
         },
 
         // eslint-disable-next-line
         updateInput: function (str) {
-            throw Error("Not Implemented!");
+            throw Error(_t("Not Implemented!"));
         },
 
         // eslint-disable-next-line
         updateShadowInput: function (str) {
-            throw Error("Not Implemented!");
+            throw Error(_t("Not Implemented!"));
         },
 
         // eslint-disable-next-line
         preventLostInputFocus: function (ev) {
-            throw Error("Not Implemented!");
+            throw Error(_t("Not Implemented!"));
         },
 
         focus: function () {
-            throw Error("Not Implemented!");
+            throw Error(_t("Not Implemented!"));
         },
 
         getUserInput: function () {
-            throw Error("Not Implemented!");
+            throw Error(_t("Not Implemented!"));
         },
 
         // eslint-disable-next-line
         isPinned: function () {
-            throw Error("Not Implemented!");
+            throw Error(_t("Not Implemented!"));
         },
 
         /* PRINT */
         flush: function () {
-            throw Error("Not Implemented!");
+            throw Error(_t("Not Implemented!"));
         },
 
         // eslint-disable-next-line
         printHTML: function (html) {
-            throw Error("Not Implemented!");
+            throw Error(_t("Not Implemented!"));
         },
 
         // eslint-disable-next-line
         print: function (msg, enl, cls) {
-            throw Error("Not Implemented!");
+            throw Error(_t("Not Implemented!"));
         },
 
         // eslint-disable-next-line
         eprint: function (msg, enl) {
-            throw Error("Not Implemented!");
+            throw Error(_t("Not Implemented!"));
         },
 
         // eslint-disable-next-line
         printCommand: function (cmd, secured = false) {
-            throw Error("Not Implemented!");
+            throw Error(_t("Not Implemented!"));
         },
 
         // eslint-disable-next-line
         printError: function (error, internal = false) {
-            throw Error("Not Implemented!");
+            throw Error(_t("Not Implemented!"));
         },
 
         // eslint-disable-next-line
         printTable: function (columns, tbody) {
-            throw Error("Not Implemented!");
+            throw Error(_t("Not Implemented!"));
         },
     });